+++ src/utils/settings.py
@@ -94,7 +94,7 @@ APPLICATION = "commix"
 DESCRIPTION_FULL = "Automated All-in-One OS Command Injection and Exploitation Tool"
 DESCRIPTION = "The command injection exploiter"
 AUTHOR  = "Anastasios Stasinopoulos"
-VERSION_NUM = "1.4.10"
+VERSION_NUM = "1.4.11"
 STABLE_VERSION = False
 if STABLE_VERSION:
   VERSION = VERSION_NUM[:3]
@@ -102,6 +102,7 @@ else:
   VERSION = VERSION_NUM[:3] + "-dev#" + VERSION_NUM[4:]
 YEAR = "2014-2016"
 AUTHOR_TWITTER = "@ancst" 
+APPLICATION_URL = "http://commixproject.com" 
 APPLICATION_TWITTER = "@commixproject" 
 
 # Inject Tag