+++ readme/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 1.4 (upcoming)
+* Updated: The Unicorn tool has been updated to version 2.3.5.
 * Added: The project's official URL has been added in the menu banner.
 * Fixed: Minor improvements in tab completion.
 * Fixed: Minor improvement in the function that checks for updates on start up.